feature. Enhance js content replacer #1. Remove data-testid attribute
